+++ pkg/thanos/statefulset.go
@@ -168,7 +168,7 @@ func makeStatefulSetSpec(tr *monitoringv1.ThanosRuler, config Config, ruleConfig
 	}
 
 	if version.GTE(semver.MustParse("0.37.0")) && tr.Spec.RuleConcurrentEval != nil && *tr.Spec.RuleConcurrentEval > 0 {
-		trCLIArgs = append(trCLIArgs, monitoringv1.Argument{Name: "rule-concurrent-evaluation", Value: strconv.FormatInt(*tr.Spec.RuleConcurrentEval, 10)})
+		trCLIArgs = append(trCLIArgs, monitoringv1.Argument{Name: "rule-concurrent-evaluation", Value: strconv.FormatInt(int64(*tr.Spec.RuleConcurrentEval), 10)})
 	}
 
 	trEnvVars := []v1.EnvVar{
